Passive player healing
The player passively heals a fifth of their total health after 10 seconds of not being hit. If the player gets hit during regen, the timer resets and the player has to wait another 10 seconds to heal again.

The time between healing is set by the variable TimeBetweenHealing. This makes the gameplay more enjoyable as it increases the survivability of the player, and was approved as a feature by Harry and Paul.