Better player inventory management
Rather than having a seperate component for EACH ITEM, one component was used for the static and skeletal meshes, being swapped out dynamically depending on the item in the player's hands. Instead of having a different model for the potions and the fireballs, one model was used and the materials were swapped out. The code is a lot cleaner now :))